requirements.txt
setuptools==68.2.2
wheel
numpy==1.26.4
gradio==3.19.1  # Add Gradio package
langchain-groq==0.1.0
langchain-community==0.0.30
pypdf==3.16.4
sentence-transformers==2.5.1
python-dotenv==1.0.1
langchain
faiss-cpu
transformers==4.38.2
huggingface_hub==0.21.3
protobuf==4.25.3
torch==2.2.1
